fix OTP code bypass

+++ app/controllers/users/sessions_controller.rb
@@ -13,13 +13,16 @@ class Users::SessionsController < Devise::SessionsController
 
   def create
     params[:user][:password] = trim_long_password(params[:user][:password])
-    self.resource = warden.authenticate!(auth_options)
+    user = User.find_for_authentication(email: params[:user][:email])
 
-    if resource&.otp_module_enabled?
+    if user&.otp_module_enabled? && user&.valid_password?(params[:user][:password])
       sign_out(resource)
-      session[:pending_user_id] = resource.id
+      session[:pending_user_id] = user.id
+      session[:remember_me] = params[:user][:remember_me]
+      self.resource = user
       render :two_factor
     else
+      self.resource = warden.authenticate!(auth_options)
       continue_sign_in(resource_name, resource)
     end
   end
@@ -28,14 +31,23 @@ class Users::SessionsController < Devise::SessionsController
     return unless session[:pending_user_id]
 
     self.resource = User.find(session[:pending_user_id])
+    return render :two_factor unless params.dig(:user, :otp_code_token).present?
 
     if Users::VerifyOtp.call(resource, params[:user][:otp_code_token])
+
+      # manually set the remember_me cookie. It's normally set by warden, but we can't use
+      # warden.authenticate!(auth_options) here because we don't have access to the user
+      # credentials (email and password). We cannot call warden.authenticate!(auth_options)
+      # before the OTP verification because then the user would be signed in directly.
+      custom_remember_me(resource) if session[:remember_me] == '1'
+
       session.delete(:pending_user_id)
-      sign_in(resource_name, resource)
-      respond_with(resource, location: after_sign_in_path_for(resource))
+      session.delete(:remember_me)
+
+      continue_sign_in(resource_name, resource)
     else
       flash.now[:alert] = t('errors.messages.bad_2fa_code')
-      render :two_factor
+      render :two_factor, status: :unprocessable_entity
     end
   end
 
@@ -43,14 +55,7 @@ class Users::SessionsController < Devise::SessionsController
 
   def continue_sign_in(resource_name, resource)
     sign_in(resource_name, resource)
-    yield resource if block_given?
-    respond_with(resource, location: after_sign_in_path_for(resource))
-  end
-
-  def abort_sign_in
-    sign_out(resource)
-    set_new_instance_variables
-    respond_with_navigational(resource) { render :new }
+    respond_with resource, location: after_sign_in_path_for(resource)
   end
 
   def set_new_instance_variables
@@ -65,5 +70,31 @@ class Users::SessionsController < Devise::SessionsController
 
   def trim_long_password(password)
     password[0...Devise.password_length.max]
+  end
+
+  def custom_remember_me(resource)
+    scope = Devise::Mapping.find_scope!(resource)
+    resource.remember_me!
+    cookies.signed[remember_key(resource, scope)] = remember_cookie_values(resource)
+  end
+
+  # from devise gem: lib/devise/controllers/rememberable.rb
+  def forget_cookie_values(resource)
+    Devise::Controllers::Rememberable.cookie_values.merge!(resource.rememberable_options)
+  end
+
+  # from devise gem: lib/devise/controllers/rememberable.rb
+  def remember_cookie_values(resource)
+    options = { httponly: true }
+    options.merge!(forget_cookie_values(resource))
+    options.merge!(
+      value: resource.class.serialize_into_cookie(resource),
+      expires: resource.remember_expires_at
+    )
+  end
+
+  # from devise gem: lib/devise/controllers/rememberable.rb
+  def remember_key(resource, scope)
+    resource.rememberable_options.fetch(:key, "remember_#{scope}_token")
   end
 end
